+++ src/main/java/com/bolghari/chattyApp/controller/rest/UserRestController.java
@@ -3,17 +3,13 @@ package com.bolghari.chattyApp.controller.rest;
 import com.bolghari.chattyApp.model.User;
 import com.bolghari.chattyApp.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.annotation.CurrentSecurityContext;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
-import java.security.Principal;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @CrossOrigin(origins = "http://localhost:8080")
@@ -35,23 +31,38 @@ public class UserRestController {
     }
 
     @GetMapping("/user/{username}")
-    @PreAuthorize("#username == principal.username")
-    public User getUser(@PathVariable("username") String username) {
-            return service.getUser(username);
-
+    @PreAuthorize("#username == principal.username or hasRole('ROLE_ADMIN')")
+    public ResponseEntity<User> getUser(@PathVariable("username") String username) {
+        User user = service.getUser(username);
+        if(user != null) {
+            return new ResponseEntity<>(service.getUser(username), HttpStatus.OK);
+        }else {
+            return new ResponseEntity("User with username: %s not found"+ username, HttpStatus.OK);
+        }
     }
 
-    //@PreAuthorize("hasAnyRole('ROLE_ADMIN')")
+
     @DeleteMapping("/delete/{id}")
-    @PreAuthorize("#id == principal.id")
-    public void deleteUser(@PathVariable("id") String id) {
+    @PreAuthorize("#id == principal.id || hasAnyRole('ROLE_ADMIN')")
+    public ResponseEntity<String>  deleteUser(@PathVariable("id") String id) {
+            Optional<User> exist = service.getUserById(id);
 
-        service.deleteUser(id);
+            if(!exist.isEmpty()) {
+                service.deleteUser(id);
+                return new ResponseEntity<>("Successfully deleted user: "+exist.get().getUsername(), HttpStatus.OK);
+            }
+            return new ResponseEntity("No user by the id found.."+id,HttpStatus.NOT_FOUND);
     }
 
 
-    @PutMapping("/edit")
-    public void updateUser() {
-
+    @PutMapping("/edit/{id}")
+    @PreAuthorize("#id == principal.id || hasAnyRole('ROLE_ADMIN')")
+    public ResponseEntity<String> updateUser(@RequestBody User user, @PathVariable("id") String id) {
+        Optional<User> exist = service.getUserById(id);
+        if(!exist.isEmpty()) {
+            service.updateUser(user, id);
+            return new ResponseEntity<>("Successfully edit user with id: "+id, HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Error, could not find the user. Please check for typo", HttpStatus.BAD_REQUEST);
     }
 }
